+++ components/layout/footer/footer-bottom.tsx
@@ -4,7 +4,7 @@ import { UiContent } from "@/types/ui-content";
 type FooterBottomProps = {
   uiContent: UiContent;
   horizontalBanner: FeaturesType["eventStyle"];
-  dimensions: FeaturesType["eventDimensions"];
+  dimensions: FeaturesType["bannerSize"];
 };
 
 export default function FooterBottom({
@@ -20,13 +20,19 @@ export default function FooterBottom({
     uiContent.companyName +
     " " +
     uiContent.footer.legal;
-    
 
   return (
-    <div className={`mt-5 border-t py-10 text-sm ${horizontalBanner == "horizontal" ? 
-  (dimensions?.dimensions == "small" ? `mb-${dimensions.small}` : 
-  dimensions?.dimensions == "medium" ? `mb-${dimensions.medium}` : 
-  "" ): ""}`}>
+    <div
+      className={`mt-5 border-t py-10 text-sm ${
+        horizontalBanner == "horizontal"
+          ? dimensions == "small"
+            ? `mb-36`
+            : dimensions == "medium"
+              ? `mb-44`
+              : ""
+          : ""
+      }`}
+    >
       <div className="mx-auto flex w-full max-w-7xl flex-col items-center gap-1 px-4 md:flex-row md:gap-0 md:px-4 min-[1320px]:px-0">
         <p>{copyrightName}</p>
         <hr className="mx-4 hidden h-4 w-[1px] border-l md:inline-block" />
@@ -43,7 +49,5 @@ export default function FooterBottom({
         </p>
       </div>
     </div>
-    
   );
-  
 }
