+++ components/upcoming-event/banner/horizontal-banner.tsx
@@ -1,14 +1,14 @@
 import { createLink } from "@/lib/create-link";
 import { LocaleDetails } from "@/types/languages";
 import { UpcomingEvent } from "@/types/upcoming-event";
-import { MainImageUpcoming } from "./main-image";
-import { ReadMoreButton } from "../read-more-button";
+import { MainImageUpcoming } from "../main-image";
+import { ReadMoreButton } from "../../read-more-button";
 import { FeaturesType } from "@/types/features";
 
 type UpcomingEventBannerCompProps = {
   upcomingEvent: UpcomingEvent;
   locale: LocaleDetails;
-  dimensions?: FeaturesType["eventDimensions"];
+  dimensions?: FeaturesType["bannerSize"];
 };
 
 export const HorizontalBanner = ({
@@ -19,8 +19,8 @@ export const HorizontalBanner = ({
   return (
     <a
       href={createLink(upcomingEvent.link, locale)}
-      className={`fixed bottom-0 right-0 ${dimensions?.dimensions == "small" ? `h-${dimensions.small}` : (dimensions?.dimensions == "medium" ? `h-${dimensions.medium}` : "")} 
-          w-screen z-50 flex items-end justify-end`}
+      className={`fixed bottom-0 right-0 ${dimensions == "small" ? `h-36` : dimensions == "medium" ? `h-44` : ""} 
+          z-50 flex w-screen items-end justify-end`}
     >
       <div className="absolute inset-0 h-full w-full">
         <MainImageUpcoming upcomingEvent={upcomingEvent} banner />
