+++ components/upcoming-event/banner/wrapper.tsx
@@ -1,0 +1,27 @@
+// `use client`;
+
+// import { useEffect, useRef } from 'react';
+
+// type WrapperProps = {
+//     children: React.ReactNode;
+// };
+
+// export function BannerWrapper({ children }: WrapperProps) {
+//     const bannerWrapperRef = useRef<HTMLDivElement | null>(null);
+
+//     useEffect(() => {
+//         const handleScroll = () => {
+//             if (window.scrollY >= 400 && bannerWrapperRef.current) {
+//                 bannerWrapperRef.current.classList.add('animate-fadeInFromBottom');
+//             }
+//         };
+
+//         window.addEventListener('scroll', handleScroll);
+
+//         return () => {
+//             window.removeEventListener('scroll', handleScroll);
+//         };
+//     }, []);
+
+//     return <div ref={bannerWrapperRef}>{children}</div>;
+// }
